미술, 한우 카테고리 수정 15

+++ src/main/java/com/moaguide/service/hanwoo/HanwooService.java
@@ -14,6 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.sql.Date;
@@ -44,7 +45,7 @@ public class HanwooService {
                 (String) result[3],  // type
                 (String) result[4],  // name
                 ((Long) result[5]),  // recruitmentPrice (Integer 그대로 사용)
-                ((Double) result[6]),  // recruitmentRate (Double 그대로 사용)
+                ((BigDecimal) result[6]).doubleValue(),  // recruitmentRate (Double 그대로 사용)
                 ((Long) result[7]),  // totalPrice
                 ((Date) result[8]).toLocalDate(),  // paymentDate
                 ((Integer) result[9])  // minInvestment
